+++ utils/responses/pagination.go
@@ -1,4 +1,4 @@
-package outputs
+package responses
 
 import (
 	"context"
