+++ utils/responses/schema.go
@@ -1,4 +1,4 @@
-package outputs
+package responses
 
 import (
 	"reflect"
